@@ -6,6 +6,7 @@ var Lang   = Y.Lang,
     APIList = Y.namespace('APIList'),
 
     classesNode    = Y.one('#api-classes'),
+    elementsNode   = Y.one('#api-elements'),
     inputNode      = Y.one('#api-filter'),
     modulesNode    = Y.one('#api-modules'),
     tabviewNode    = Y.one('#api-tabview'),
@@ -97,7 +98,9 @@ tabview.get('panelNode').all('a').each(function (link) {
 
 // -- Private Functions --------------------------------------------------------
 function getFilterResultNode() {
-    return filter.get('queryType') === 'classes' ? classesNode : modulesNode;
+    var queryType = filter.get('queryType');
+    return queryType === 'classes' ? classesNode
+            : queryType === 'elements' ? elementsNode : modulesNode;
 }
 
 // -- Event Handlers -----------------------------------------------------------
@@ -105,7 +108,7 @@ function onFilterResults(e) {
     var frag         = Y.one(Y.config.doc.createDocumentFragment()),
         resultNode   = getFilterResultNode(),
         typePlural   = filter.get('queryType'),
-        typeSingular = typePlural === 'classes' ? 'class' : 'module';
+        typeSingular = typePlural === 'classes' ? 'class' : typePlural === 'elements' ? 'element' : 'module';
 
     if (e.results.length) {
         YArray.each(e.results, function (result) {
@@ -181,6 +184,7 @@ function onTabSelectionChange(e) {
     };
 
     switch (name) {
+    case 'elements':// fallthru
     case 'classes': // fallthru
     case 'modules':
         filter.setAttrs({
